### 2025-12-09 #9396 @EwoutH
@aslan-ng looks like an interesting paper and library, I'm definitely going to learn something from it!

Before we go into the checklist process, I'm curious, did you consider extending or modifying an existing ABM library?

CC @quaquel, [this short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.09396/joss.09396/10.21105.joss.09396.pdf) is worth a read, and [the library](https://github.com/cmudrc/pied-piper) API worth taking a look at.
