### 2025-12-10 #9602 @editorialbot
[REVIEW]: PiperABM: A Python Library for Resilience-Based Agent Modeling
**Submitting author:** @aslan-ng (<a href="http://orcid.org/0009-0004-3387-4502">Aslan Noorghasemi</a>)
**Repository:** https://github.com/cmudrc/pied-piper
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
