### 2025-12-15 a85064d
Fix "infrastrcuture" typo in step-by-step documentation
Replace the use of `infrastrcuture` with the correctly spelled `infrastructure`.
