### 2025-12-15 #9602 @EwoutH
I started looking into the repo. It's nicely designed and uses interesting concepts.

I left some more detailed feedback in the repo about i.e. API design, but none of that is hard required for the paper.

> Did you consider extending or modifying an existing ABM library?

After looking into PiperABM a bit more, I'm even more curious about this!
