### 2026-01-05 0dfd661
Contribution guide added to the docs and easier testing procedure implemented.
