### 2026-01-05 eb29d26
Exception added for baked scenario fails.
