### 2026-01-05 610bb6a
Resource class added.
