+++ piperabm/society/query/add.py
@@ -2,6 +2,7 @@ import numpy as np
 from copy import deepcopy
 
 from piperabm.exceptions import ModelNotBakedError
+from piperabm.resource import Resource
 from piperabm.society.actions.action_queue import ActionQueue
 from piperabm.society.info import *
 
@@ -35,16 +36,70 @@ class Add:
         id: int = None,
         name: str = "",
         socioeconomic_status: float = 1,
-        resources: dict = {
-            "food": 10,
-            "water": 10,
-            "energy": 10,
-        },
-        enough_resources: dict = None,
+        resources: dict | Resource = Resource(),
+        enough_resources: dict | Resource | None = None,
         balance: float = 0,
     ):
         """
-        Add agent node
+        Add an agent node to the society network.
+
+        Agents are represented as nodes in the society graph. Each agent is assigned to
+        a home node in the infrastructure and initialized with resources and a balance.
+        When an agent is added, family and neighbor relationships may be created
+        automatically based on the agent's home assignment.
+
+        Parameters
+        ----------
+        home_id : int, optional
+            ID of the home node that the agent belongs to. If ``None``, a home is chosen
+            at random from existing infrastructure homes.
+
+        id : int, optional
+            Unique identifier of the agent node. If ``None``, a unique ID is automatically
+            generated. If the provided ID already exists, a new unique ID is generated
+            instead (the existing agent is not overwritten).
+
+        name : str, optional
+            Optional human-readable name for the agent.
+
+        socioeconomic_status : float, optional
+            Socioeconomic status of the agent (used by the decision-making and/or
+            behavior models). Defaults to ``1``.
+
+        resources : dict or Resource, optional
+            Initial resource inventory for the agent. This can be provided either as a
+            dictionary (e.g., ``{'food': 10, 'water': 10, 'energy': 10}``) or as a
+            :class:`piperabm.Resource` instance. If a ``Resource`` object is provided,
+            it is converted internally to a dictionary before being attached to the
+            underlying graph representation.
+
+        enough_resources : dict or Resource or None, optional
+            The per-resource "enough" thresholds used by the agent's satisfaction/utility
+            model. If provided as a ``Resource`` object, it is converted internally to a
+            dictionary. If ``None``, the system initializes the threshold for each
+            resource to match the corresponding initial ``resources`` value.
+
+        balance : float, optional
+            Initial monetary balance of the agent. Defaults to ``0``.
+
+        Returns
+        -------
+        int
+            The ID of the newly added agent node.
+
+        Raises
+        ------
+        ModelNotBakedError
+            If the infrastructure has not been baked. Agents require a finalized
+            infrastructure network (via :meth:`~piperabm.Model.bake`) to ensure
+            physically consistent routing and access edges.
+
+        Notes
+        -----
+        Internally, agent state (including resources and thresholds) is stored as plain
+        attributes on the NetworkX society graph. The :class:`piperabm.Resource` class
+        is provided as an optional convenience wrapper for validation/readability and
+        does not change the internal representation.
         """
         if self.infrastructure.baked is False:
             raise ModelNotBakedError("Model is not baked.")
@@ -56,6 +111,10 @@ class Add:
             homes_id = self.infrastructure.homes
             home_id = int(np.random.choice(homes_id))
         pos = self.infrastructure.get_pos(id=home_id)
+        if isinstance(resources, Resource):
+            resources = resources.serialize()
+        if isinstance(enough_resources, Resource):
+            enough_resources = enough_resources.serialize()
         resource_kwargs = {}
         if enough_resources is None:
             enough_resources = {}
@@ -103,7 +162,34 @@ class Add:
 
     def add_family(self, id_1: int, id_2: int):
         """
-        Add family edge
[…]
