+++ docs/build/html/_sources/api/measurement.rst.txt
@@ -1,4 +1,4 @@
-piperabm.Measurement
+Measurement
 ======================================
 
 .. autoclass:: piperabm.model.Measurement
