+++ docs/build/html/_sources/api/infrastructure.rst.txt
@@ -1,4 +1,4 @@
-piperabm.Infrastructure
+Infrastructure
 ======================================
 
 .. autoclass:: piperabm.infrastructure.Infrastructure
