+++ docs/build/html/_sources/index.rst.txt
@@ -32,6 +32,7 @@ This is the documentation for PiperABM, a Python-based agent-based modeling fram
    api/measurement
    api/infrastructure
    api/society
+   api/resource
    api/projection
    api/mesh
 
