### 2026-01-06 980626e
More tests for resource class.
