+++ piperabm/society/query/set.py
@@ -1,4 +1,5 @@
 from piperabm.tools.nx_query import NxSet
+from piperabm.resource import Resource
 
 
 class Set(NxSet):
@@ -24,10 +25,16 @@ class Set(NxSet):
             self.set_node_attribute(id=id, attribute="alive", value=False)
         self.set_node_attribute(id=id, attribute=name, value=value)
 
-    def set_resources(self, id: int, values: dict) -> None:
+    def set_resources(self, id: int, values: dict | Resource) -> None:
         """
-        Set agent resources values
+        Set agent resources values.
         """
+        if isinstance(values, Resource):
+            values = dict(values)
+            for name in values:
+                # Delete the key if the value is 0
+                if values[name] == 0:
+                    del values[name]
         for name in values:
             self.set_resource(id=id, name=name, value=values[name])
 
