### 2026-01-06 a731f25
Pied-piper image links in both docs and readme updated.
