+++ docs/build/html/_sources/api/latlong_xy.rst.txt
@@ -1,7 +1,0 @@
-piperabm.tools.coordinate.projection
-======================================
-
-.. automodule:: piperabm.tools.coordinate.projection.latlong_xy
-   :members:
-   :undoc-members:
-   :show-inheritance:
