### 2026-01-06 71526a6
Mor tests for resource class.
