### 2026-01-06 2ffa166
Resources class guide improved.
