### 2026-01-08 68dd7db
Custom degradation and decision-making classes are served in a more pythonic way.
