+++ piperabm/society/decision_making.py
@@ -12,10 +12,10 @@ class DecisionMaking:
         Return preassumed destinations (which are market nodes here)
         """
         result = []
-        nodes = self.infrastructure.markets
+        nodes = self.society.infrastructure.markets
         for node in nodes:
-            path_exists = self.infrastructure.has_path(
-                id_start=self.get_current_node(id=agent_id), id_end=node
+            path_exists = self.society.infrastructure.has_path(
+                id_start=self.society.get_current_node(id=agent_id), id_end=node
             )
             if path_exists is True:
                 result.append(
@@ -35,21 +35,21 @@ class DecisionMaking:
         result = []
         nodes = set()
         # Friends
-        friends = self.ego(id=agent_id, type="friend")
+        friends = self.society.ego(id=agent_id, type="friend")
         friends_homes = []
         for friend in friends:
-            friends_homes.append(self.get_home_id(id=friend))
+            friends_homes.append(self.society.get_home_id(id=friend))
         nodes |= set(friends_homes)
         # Neighbors
-        neighbors = self.ego(id=agent_id, type="neighbor")
+        neighbors = self.society.ego(id=agent_id, type="neighbor")
         neighbors_homes = []
         for neighbor in neighbors:
-            neighbors_homes.append(self.get_home_id(id=neighbor))
+            neighbors_homes.append(self.society.get_home_id(id=neighbor))
         nodes |= set(neighbors_homes)
         nodes = list(nodes)
         for node in nodes:
-            path_exists = self.infrastructure.has_path(
-                id_start=self.get_current_node(id=agent_id), id_end=node
+            path_exists = self.society.infrastructure.has_path(
+                id_start=self.society.get_current_node(id=agent_id), id_end=node
             )
             if path_exists is True:
                 result.append(
@@ -71,27 +71,27 @@ class DecisionMaking:
         # Calculate the estimated amount of fuel required
         travel_duration = self.estimated_duration(agent_id, destination_id)
         fuel_resources = {}
-        for name in self.resource_names:
+        for name in self.society.resource_names:
             fuel_resources[name] = (
-                self.transportation_resource_rates[name] * travel_duration
+                self.society.transportation_resource_rates[name] * travel_duration
             )
         # Calculate the value of required fuel
         fuel_possible = True
-        for name in self.resource_names:
-            if fuel_resources[name] > self.get_resource(id=agent_id, name=name):
+        for name in self.society.resource_names:
+            if fuel_resources[name] > self.society.get_resource(id=agent_id, name=name):
                 fuel_possible = False
         if fuel_possible is True:
             total_fuel_value = 0
-            for name in self.resource_names:
-                fuel_value = fuel_resources[name] * self.prices[name]
+            for name in self.society.resource_names:
+                fuel_value = fuel_resources[name] * self.society.prices[name]
                 total_fuel_value += fuel_value
         else:
             total_fuel_value = SYMBOLS["inf"]
         # Calculate the value of resources there
-        resources_there = self.resources_in(node_id=destination_id, is_market=is_market)
+        resources_there = self.society.resources_in(node_id=destination_id, is_market=is_market)
         total_value_there = 0
-        for name in self.resource_names:
-            total_value_there += resources_there[name] * self.prices[name]
+        for name in self.society.resource_names:
+            total_value_there += resources_there[name] * self.society.prices[name]
         # Calculate score
         score = total_value_there - total_fuel_value
         return score
@@ -100,8 +100,8 @@ class DecisionMaking:
         """
         Estimated distance between agent and destination
         """
-        return self.infrastructure.heuristic_paths.estimated_distance(
-            id_start=self.get_current_node(id=agent_id), id_end=destination_id
+        return self.society.infrastructure.heuristic_paths.estimated_distance(
[…]
