+++ piperabm/model/__init__.py
@@ -1,4 +1,2 @@
 from .model import Model
 from .measurement import Measurement
-
-# from .grid_world import grid_world
