### 2026-01-08 f2f8c53
Example updated with more relevant info.
