### 2026-01-08 bc4b1b2
New class injection method is reflected in the docs.
