### 2026-01-08 #9602 @EwoutH
Quick update: I provided some feedback directly in their repo over the past week, and got great reactions from @aslan-ng on it.

Before continuing with my checklist, I would like a reply to this question:

> Did you consider extending or modifying an existing ABM library? I think the need for a specialized solution is clear in the paper, but not yet why that needed to be a completely new solution.
