### 2026-01-08 b104b5f
Animation guide added to docs.
