+++ docs/source/conf.py
@@ -13,7 +13,7 @@ sys.path.insert(0, os.path.abspath("../.."))
 project = "PiperABM"
 copyright = "2025, Aslan Noorghasemi, Christopher McComb"
 author = "Aslan Noorghasemi, Christopher McComb"
-release = "0.1.1"
+release = "0.1.2"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -33,5 +33,5 @@ exclude_patterns = []
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = "alabaster"
+html_theme = "furo" # Default: "alabaster"
 html_static_path = ["_static"]
