### 2026-01-11 106f40a
Joss badge added to readme.
