### 2026-01-11 ba0a855
Organizing docs and readme.
