+++ docs/source/index.rst
@@ -18,20 +18,36 @@ This is the documentation for PiperABM, a Python-based agent-based modeling fram
    
    Illustration by Kate Greenaway, originally published in Robert Browning’s *The Pied Piper of Hamelin*.
 
-.. toctree::
-   :maxdepth: 1
-   :caption: Installation
+Getting Started
+---------------
 
-   installation
+- Start with :doc:`usage/installation`
+- Follow the :doc:`usage/step-by-step` tutorial
+- Explore practical examples in the
+  `examples directory <https://github.com/cmudrc/pied-piper/tree/main/examples>`_
+
+Project Links
+-------------
+
+- **Source code (GitHub):** https://github.com/cmudrc/pied-piper
+- **Python package (PyPI):** https://pypi.org/project/piperabm/
+- **JOSS paper:** https://joss.theoj.org/papers/482e67b36285c6a9e5c00b15cc26607f
+
+If you use PiperABM in academic work, please cite the accompanying JOSS paper (currently under review).
+
 
 .. toctree::
+   :hidden:
    :maxdepth: 2
    :caption: Usage Guide
 
-   step-by-step
-   satellite
+   usage/installation
+   usage/step-by-step
+   usage/satellite
+   usage/backend
 
 .. toctree::
+   :hidden:
    :maxdepth: 1
    :caption: API Reference
       
@@ -39,13 +55,13 @@ This is the documentation for PiperABM, a Python-based agent-based modeling fram
    api/measurement
    api/infrastructure
    api/society
-   api/backend
    api/resource
    api/projection
    api/mesh
 
 .. toctree::
+   :hidden:
    :maxdepth: 1
    :caption: Contributing
 
-   contributing
+   contributing/contributing
