### 2026-01-11 e917f47
More test for measurement
