### 2026-01-13 fdd32c8
Release v0.2.0: explicit behavior registration and docs updates
