+++ piperabm/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.1.2"
+__version__ = "0.2.0"
 
 from .model import Model
 from .model import Measurement
