+++ docs/source/conf.py
@@ -13,7 +13,7 @@ sys.path.insert(0, os.path.abspath("../.."))
 project = "PiperABM"
 copyright = "2025, Aslan Noorghasemi, Christopher McComb"
 author = "Aslan Noorghasemi, Christopher McComb"
-release = "0.1.2"
+release = "0.2.0"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
