+++ requirements.txt
@@ -1,4 +1,4 @@
-keepdelta==0.2.0
+keepdelta
 matplotlib==3.10.1
 networkx==3.4.2
 numpy==2.2.5
