+++ pyproject.toml
@@ -10,7 +10,7 @@ include = ["piperabm", "piperabm.*"]
 
 [project]
 name = "piperabm"
-version = "0.1.2"
+version = "0.2.0"
 description = "An Agent-Based Model Framework for Simulating Resilience in Python"
 readme = "README.md"
 license = {text = "MIT"}
@@ -20,11 +20,11 @@ authors = [
     { name = "Christopher McComb", email = "ccm@cmu.edu" },
 ]
 dependencies = [
-    "keepdelta==0.2.0",
-    "matplotlib",
-    "networkx",
-    "numpy",
-    "scipy",
+    "keepdelta",
+    "matplotlib==3.10.1",
+    "networkx==3.4.2",
+    "numpy==2.2.5",
+    "scipy==1.15.2",
 ]
 requires-python = ">=3.10"
 classifiers = [
