### 2026-01-30 #9602 @nikos-salamanos
**My Review Summary:**
**Installation:** Works on Python 3.12. Note: Installation failed on Python 3.10
**Functionality:** Core functionality runs as described.
**Tests:** The automated test suite passes successfully.
**Documentation:** The “Step-by-Step” tutorial fails to run and requires updates to align with the current API.
**Paper:** The “Statement of Need” requires clarification regarding the scope of the supported behavioral mechanisms.

**Detailed Findings:**

**1. Issue: The “Statement of Need” needs an update on the scope of the supported behavioral mechanisms**
The current implementation of PiperABM provides a baseline for agent decision-making and trading mechanisms. More complex behavioral dynamics (e.g., alternative bargaining models or weighted social interactions) are not supported by default (via the public API) and will require user customization. Clarifying these limitations in the paper and documentation would help users manage their expectations regarding the library. The library is best suited for studying system-level redistribution and macro-level outcomes under cooperative trading assumptions (Nash Bargaining).

Specifically:
The paper states (lines 56–58) that
“...with these pieces, researchers can quickly pose hypotheses, run ‘what-if’ scenarios…”.
This is accurate primarily for parameter-driven and scenario-based research questions, since some core behavioral mechanisms are fixed in the current implementation.

For example, trading between agents is implemented using Nash bargaining, a cooperative game-theoretic mechanism. This is a strong behavioral assumption. I did not find a clearly described, supported way for users to replace the built-in bargaining mechanism without extending internal components. Moreover, trades are resolved via a centralized optimization (e.g., trade_solver.py, nash_bargaining.py) rather than through decentralized, agent-to-agent interaction.

Similarly, social relations (family / neighbors / friends) are represented structurally as graph edges, but are not behaviorally parameterized by default. Even if we define social tie strength via edge weights, this will not affect the default agents’ behavior. In short, modeling custom preferential attachments among agents (based on strong and weak ties) will require user-implemented customization, as there is no built-in mechanism for this.

Clarifying the scope of supported behavioral assumptions and which components are configurable versus fixed would help users better understand which kinds of hypotheses the library can support.

**2. Issue: Installation does not support Python 3.10**
The documentation states that PiperABM is officially supported on Python 3.10 to 3.13. However, both the released version (pip install … ) and the repository version fail in a Python 3.10 environment.
By just importing the library, I got a SyntaxError (self.G.edges[*ids] in piperabm/tools/nx_query/get.py). This syntax requires Python 3.11.
Please update the documentation or the code accordingly.

**3. Issue: “The Step-by-Step” tutorial fails**
While running the step-by-step tutorial on the repository version (Python 3.12 env), it fails during model.infrastructure.bake(...) with AttributeError: edges_closer_than.
Although edges_closer_than is referenced in piperabm/infrastructure/grammar/rules/rule_3.py , there is no corresponding method anywhere in the piperabm source. Please implement this method or update Rule 3 to call the correct function.
