+++ requirements.txt
@@ -1,5 +1,5 @@
 keepdelta
-matplotlib==3.10.1
-networkx==3.4.2
-numpy==2.2.5
-scipy==1.15.2
+matplotlib
+networkx
+numpy
+scipy
