### 2026-04-12 bfc6e0c
drop python 3.10 support
