+++ docs/build/html/_sources/usage/installation.rst.txt
@@ -11,7 +11,7 @@ Install the package using pip:
 Python version requirements
 ---------------------------
 
-PiperABM is tested and officially supported on **Python 3.10 through 3.13**.
+PiperABM is tested and officially supported on **Python 3.11 through 3.13**.
 
 While the package may work on older Python versions with compatible dependencies,
 these environments are not tested and are not officially supported.
