+++ pyproject.toml
@@ -26,7 +26,7 @@ dependencies = [
     "numpy",
     "scipy",
 ]
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
