### 2026-04-12 4d51d3c
smoke test added.
