### 2026-04-12 666c265
edge_closer_than method added.
