### 2026-04-12 30ab614
Address reviewer comments: clarify behavioral scope
