+++ paper/paper.bib
@@ -410,3 +410,42 @@
     month = feb,
 	note = {[Accessed 21-09-2025]},
 }
+
+@article{Noorghasemi2025-rf,
+  title    = "Agent-based modeling for the evaluation of community resilience
+              in silico",
+  author   = "Noorghasemi, Aslan and Xiao, Yinshuang and Sha, Zhenghui and
+              Xiao, Ming and McComb, Christopher",
+  abstract = "not-yet-known not-yet-known not-yet-known unknown Civil
+              infrastructure is a system through which humans inter- act with
+              each other and their environment, making it essen- tial for
+              community well-being. Such systems, however, de- grade over time
+              and become less usable, either due to pro- gressive wear or
+              abrupt damage. Therefore, it becomes nec- essary to allocate
+              resources to maintain the infrastructure and thereby ensure the
+              community's well-being. However, the relationship between
+              infrastructure condition and com- munity well-being is unclear
+              and challenging to quantify. This is especially true in locations
+              that are impacted by climate change or other significant forcing
+              factors. Current compu- tational models often struggle to predict
+              cascading nonlin- ear events that occur during system failures.
+              This paper in- troduces a computational method to quantitatively
+              analyze the impact of infrastructure on community well-being,
+              com- bining agent-based modeling (ABM) and network robustness
+              measurement. We specifically examine the town of Utqiaġvik,
+              Alaska, where permafrost thaw due to global warming threat- ens a
+              substantial amount of infrastructure. The results show that the
+              breakdown of critical infrastructure progressively weak- ens
+              community access to essential resources, leading to much lower
+              robustness. Additionally, our simulations indicate that smaller
+              household sizes and redundant infrastructure design prove
+              beneficial for sustaining resource accessibility and fos- tering
+              close social connections within the community. These insights
+              offer valuable guidance for understanding the com- plex systems
+              interplay among communities, infrastructure, and the environment,
+              thereby informing strategies to build more sustainable and
+              resilient systems in remote areas such as Utqiaġvik.",
+  journal  = "Authorea Inc.",
+  month    =  nov,
+  year     =  2025
+}
