+++ paper/paper.md
@@ -38,7 +38,7 @@ _PiperABM_ is an open-source Python library for building agent-based simulations
 
 # Statement of Need
 
-Agent-based models (ABMs) represent systems “from the bottom up” using heterogeneous, interacting agents whose local rules and behaviors generate macro-level patterns macro-level patterns [@epstein1999; @epstein_axtell_1996; @schelling_2006]. ABM has also been highlighted as a framework that engages diverse stakeholders, enables the forecasting of equitable futures, and empowers policy [@thearcticinstituteAgentsArctic; @Noorghasemi2025]. Over the several decades, ABM has matured from theoretical demonstrations to a practical research method supported by accessible tooling and the scientific Python ecosystem [@Mesa_3]. This paradigm is well-suited to infrastructure and community resilience, particularly within the Food–Energy–Water (FEW) nexus, where access to essential resources depends on both physical networks and human decisions, and where nonlinear feedbacks, thresholds, and cascades are common [@Xue2024; @Zhang2019; @CansinoLoeza2022].
+Agent-based models (ABMs) represent systems “from the bottom up” using heterogeneous, interacting agents whose local rules and behaviors generate macro-level patterns [@epstein1999; @epstein_axtell_1996; @schelling_2006]. ABM has also been highlighted as a framework that engages diverse stakeholders, enables the forecasting of equitable futures, and empowers policy [@thearcticinstituteAgentsArctic; @Noorghasemi2025]. Over recent decades, ABM has matured into a practical research method supported by the scientific Python ecosystem [@Mesa_3]. This paradigm is well-suited to infrastructure and community resilience, particularly within the Food–Energy–Water (FEW) nexus, where access to essential resources depends on both physical networks and human decisions, and where nonlinear feedbacks, thresholds, and cascades are common [@Xue2024; @Zhang2019; @Noorghasemi2025-rf; @CansinoLoeza2022].
 
 Resilience studies pose several practical modeling needs that are not well served by ad hoc scripts or purely equation-based approaches:
 
@@ -48,13 +48,13 @@ Resilience studies pose several practical modeling needs that are not well serve
 - **Ensemble experiments and reproducibility:** Parameter sweeps, stochastic replications, and uncertainty and sensitivity analysis require clear, standardized reporting and careful data handling; community guidelines for documenting agent-based models and recent work on calibration and sensitivity analysis support replicable workflows [@GrimmODD2020; @McCulloch2022; @Razavi2021].
 - **Custom metrics and analysis pipelines:** Familiar data structures expose model state for domain-specific metrics, such as spatiotemporal accessibility under disruption [@Tariverdi2023; @Enderami2024].
 
-PiperABM is built to meet these needs. It offers simple building blocks in pure Python: a society network comprised of agents and their relationships; an infrastructure network that can wear down over time; tools to load real maps; built‑in measures such as accessibility and travel distance; and optional animation for evaluating face validity. With these pieces, researchers can quickly pose hypotheses, run “what‑if” scenarios, and study how people’s choices and the infrastructure around them influence one another and analyze socio-technical feedbacks.
+PiperABM is built to meet these needs, providing a baseline set of behavioral mechanisms. It offers simple building blocks in pure Python: a society network comprised of agents and their relationships; an infrastructure network that can wear down over time; tools to load real maps; built‑in measures such as accessibility and travel distance; and optional animation for evaluating face validity. With these pieces, researchers can quickly pose parameter-driven hypotheses, run “what-if” scenarios, and study how people’s choices and infrastructure co-evolve under the current behavioral assumptions, while more advanced behavioral mechanisms can be introduced through user-defined extensions.
 
[…]
