+++ paper/paper.md
@@ -54,7 +54,7 @@ Intended users are researchers and practitioners in infrastructure and community
 
 # Model Overview
 
-At each time step, agents follow OODA-style decision logic—observing current conditions, orienting, deciding, and acting [@brehmer_dynamic_ooda; @Johnson02012023_ooda]. Agents choose actions such as where to travel, whom to meet, and whether to trade. Routing between locations uses an A* search on the evolving network to approximate human‑like paths [@foead_systematic_2021]. Trades between agents are resolved using a Nash bargaining formulation [@nash_bargaining_1950], representing a cooperative, centralized mechanism in the current implementation. Infrastructure elements degrade due to both exogenous aging and endogenous usage by agents, increasing effective travel cost and feeding back into subsequent decisions. Simulation results are stored using human‑readable delta differencing (KeepDelta), which is compact and easy to debug [@Noorghasemi_KeepDelta_A_Python_2025].
+At each time step, agents follow OODA-style decision logic: observing current conditions, orienting, deciding, and acting [@brehmer_dynamic_ooda; @Johnson02012023_ooda]. Agents choose actions such as where to travel, whom to meet, and whether to trade. Routing between locations uses an A* search on the evolving network to approximate human‑like paths [@foead_systematic_2021]. Trades between agents are resolved using a Nash bargaining formulation [@nash_bargaining_1950], representing a cooperative, centralized mechanism in the current implementation. Infrastructure elements degrade due to both exogenous aging and endogenous usage by agents, increasing effective travel cost and feeding back into subsequent decisions. Simulation results are stored using human‑readable delta differencing (KeepDelta), which is compact and easy to debug [@Noorghasemi_KeepDelta_A_Python_2025].
 
 ![PiperABM overview: The computational model emulates the relation between the elements of infrastructure and social networks.](./assets/interconnected.png){#fig:overview}
 *Figure 1.* The computational model of PiperABM emulates the relation between the elements of infrastructure and social networks.
