### 2026-04-14 #9602 @EwoutH
Glad to see this moving along. I will try to review further later this week.
